requirements.txt
streamlit==1.32.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
nltk==3.8.1
google-generativeai==0.7.3
lxml==4.9.3
-e .